requirements.txt
djangorestframework
psycopg2
django-simple-history